Thijs changed the CAN0_initialize() function for different CAN bitrate. It is still not working in the normal and listen-all-message mode.
Please take a look at this reminder in documentation:
The CAN module will check for bus idle condition before entering the Normal Operation mode.
If the CAN module is not connected to the bus or a transceiver, the CAN module will not enter Normal Operation mode. This is because the CAN module will always detect a dominant state at its receive pin.

When CAN module is set to normal operation mode,
the C1INT register bit IVRIF = 1, CERRIF = 1 and MODIF = 1.

But when the CAN module is set to loopback mode, they are 0!
CAN bus can not go into normal operation.

Zhiyuan
2018-10-01 14:11